refactor(SwingSet): split storeGC tests
