+++ packages/SwingSet/test/stores/test-storeGC/test-scalar-store-kind.js
@@ -1,0 +1,26 @@
+// eslint-disable-next-line import/order
+import { test } from '../../../tools/prepare-test-env-ava.js';
+
+import { setupTestLiveslots } from '../../liveslots-helpers.js';
+import { buildRootObject, mapRef } from './gc-helpers.js';
+
+// These tests follow the model described in
+// ../virtualObjects/test-virtualObjectGC.js
+
+// NOTE: these tests must be run serially, since they share a heap and garbage
+// collection during one test can interfere with the deterministic behavior of a
+// different test.
+
+// prettier-ignore
+test.serial('assert known scalarMapStore ID', async t => {
+  // The KindID for scalarMapStore is referenced by many of these
+  // tests (it is baked into our `mapRef()` function), and it might
+  // change if new IDs are allocated before the collection types are
+  // registered. Check it explicity here. If this test fails, consider
+  // updating `mapRef()` to use the new value.
+
+  const { testHooks } = await setupTestLiveslots(t, buildRootObject, 'bob', true);
+  const id = testHooks.obtainStoreKindID('scalarMapStore');
+  t.is(id, 2);
+  t.is(mapRef('INDEX'), 'o+2/INDEX');
+});
